Refine modal style when bottom, use indicatorInset to avoid the home indicator
